pyproject.toml
[project]
name = "hitl-cli"
version = "0.1.0"
description = "Command-line interface for Human-in-the-Loop services"
license = {text = "MIT"}
authors = [
    {name = "HITL CLI Team", email = "team@example.com"},
]
dependencies = [
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "typer>=0.16.0",
    "fastmcp>=0.3.0",
    "pyjwt>=2.9.0",
]
requires-python = ">=3.8"

[project.scripts]
hitl-cli = "hitl_cli.main:main"
hitl = "hitl_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]
